lab 3 done

+++ machine_learning/Fitness/Fitness/src/fitness/Evaluable.java
@@ -13,7 +13,7 @@ public interface Evaluable {
     public static final int L=100;  // the length of the chromosome
     
     public byte[] getDNA();
-    public int getFitness();
+    public long getFitness();
     public void setFitness(int fitness);
     public Evaluable myClone();
